moved scheme-specific vectorized ops to the scheme packages
