+++ ckks/advanced/homomorphic_encoding_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/tuneinsight/lattigo/v3/ckks"
-	"github.com/tuneinsight/lattigo/v3/ring"
 	"github.com/tuneinsight/lattigo/v3/rlwe"
 	"github.com/tuneinsight/lattigo/v3/utils"
 )
@@ -176,7 +175,7 @@ func testCoeffsToSlots(params ckks.Parameters, t *testing.T) {
 		}
 
 		// Applies bit-reverse on the original complex vector
-		ring.SliceBitReverseInPlaceComplex128(values, params.Slots())
+		ckks.SliceBitReverseInPlaceComplex128(values, params.Slots())
 
 		// Maps to a float vector
 		// Add gaps if sparse packing
@@ -331,7 +330,7 @@ func testSlotsToCoeffs(params ckks.Parameters, t *testing.T) {
 		}
 
 		// Result is bit-reversed, so applies the bit-reverse permutation on the reference vector
-		ring.SliceBitReverseInPlaceComplex128(valuesReal, params.Slots())
+		ckks.SliceBitReverseInPlaceComplex128(valuesReal, params.Slots())
 
 		verifyTestVectors(params, encoder, decryptor, valuesReal, valuesTest, params.LogSlots(), 0, t)
 	})
