+++ ckks/utils.go
@@ -249,6 +249,50 @@ func scaleUpVecExactBigFloat(values []*big.Float, scale float64, moduli []uint64
 	}
 }
 
+// SliceBitReverseInPlaceComplex128 applies an in-place bit-reverse permuation on the input slice.
+func SliceBitReverseInPlaceComplex128(slice []complex128, N int) {
+
+	var bit, j int
+
+	for i := 1; i < N; i++ {
+
+		bit = N >> 1
+
+		for j >= bit {
+			j -= bit
+			bit >>= 1
+		}
+
+		j += bit
+
+		if i < j {
+			slice[i], slice[j] = slice[j], slice[i]
+		}
+	}
+}
+
+// SliceBitReverseInPlaceRingComplex applies an in-place bit-reverse permuation on the input slice.
+func SliceBitReverseInPlaceRingComplex(slice []*ring.Complex, N int) {
+
+	var bit, j int
+
+	for i := 1; i < N; i++ {
+
+		bit = N >> 1
+
+		for j >= bit {
+			j -= bit
+			bit >>= 1
+		}
+
+		j += bit
+
+		if i < j {
+			slice[i], slice[j] = slice[j], slice[i]
+		}
+	}
+}
+
 // Divides x by n^2, returns a float
 func scaleDown(coeff *big.Int, n float64) (x float64) {
